refactor(billing): break checkout modal into checkout/ folder
Split checkout.vue into index.vue plus colocated payment-status,
success-view, and checkout-footer subcomponents, with all non-UI logic
(Stripe session, success state, sfx) moved into a colocated
use-checkout.ts composable. Update the four call sites that import the
modal component.
