+++ src/components/modals/checkout/success-view.vue
@@ -1,0 +1,21 @@
+<script setup lang="ts">
+import { useI18n } from 'vue-i18n'
+import UiIcon from '@/components/ui-kit/icon.vue'
+
+const { t } = useI18n()
+</script>
+
+<template>
+  <div
+    data-testid="checkout__success"
+    class="flex flex-col items-center gap-3 px-(--sheet-px) py-10 text-center"
+  >
+    <ui-icon src="check" class="w-12 h-12 text-(--theme-primary)" />
+    <h2 data-testid="checkout__success-heading" class="text-2xl text-brown-700">
+      {{ t('billing.checkout.success-heading') }}
+    </h2>
+    <p data-testid="checkout__success-message" class="text-brown-500">
+      {{ t('billing.checkout.success-message') }}
+    </p>
+  </div>
+</template>
