+++ src/composables/member/subscription-actions.ts
@@ -3,7 +3,7 @@ import { useCancelSubscriptionMutation, useResumeSubscriptionMutation } from '@/
 import { useAlert } from '@/composables/alert'
 import { useToast } from '@/composables/toast'
 import { useModal } from '@/composables/modal'
-import Checkout from '@/components/modals/checkout.vue'
+import Checkout from '@/components/modals/checkout/index.vue'
 
 /**
  * Subscription lifecycle orchestrators for the billing plan section: upgrade a
